Fix TIF handling when mktdata index is Date
Calling as.Date on an xts object does not work, because as.Date.default
is dispatched and doesn't know how to handle xts objects. Use coredata
to extract the TIF string from the xts object before calling as.Date.

Address Brian's TODO FIXME to convert timestamp to POSIXct before
adding the numeric time.in.force (documented to be in seconds).


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/blotter@1712 edb9625f-4e0d-4859-8d74-9fd3b1da38cb