Make explicit that getOrders works with one symbol
getOrders will throw an error if 'symbol' is a vector and not in the
same order as the symbols in the portfolio. If that error is avoided
because the symbols are in the same order, then there will be an error
when we try to subset orderbook[[portfolio]] by a vector, since [[ can
only select a single element.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/blotter@1746 edb9625f-4e0d-4859-8d74-9fd3b1da38cb